addedexceptions

+++ src/main/java/com/example/dentalclinic/service/impl/AppointmentServiceImpl.java
@@ -2,7 +2,11 @@ package com.example.dentalclinic.service.impl;
 
 import com.example.dentalclinic.entity.Appointment;
 import com.example.dentalclinic.entity.Patient;
+import com.example.dentalclinic.exceptions.BadRequestException;
+import com.example.dentalclinic.exceptions.ResourceNotFoundException;
 import com.example.dentalclinic.repository.AppointmentRepository;
+import com.example.dentalclinic.repository.DentistRepository;
+import com.example.dentalclinic.repository.PatientRepository;
 import com.example.dentalclinic.service.AppointmentService;
 import lombok.RequiredArgsConstructor;
 
@@ -16,12 +20,20 @@ import java.util.Optional;
 public class AppointmentServiceImpl implements AppointmentService {
 
     private final AppointmentRepository appointmentRepository;
+    private final PatientRepository patientRepository;
+    private final DentistRepository dentistRepository;
 
 
     @Override
-    public Appointment save(Appointment appointment) {
+    public Appointment save(Appointment appointment) throws BadRequestException { //TODO: chequear
 
-        return appointmentRepository.save(appointment);
+        if(patientRepository.findById(appointment.getPatient().getId()).isPresent() &&
+                dentistRepository.findById(appointment.getDentist().getId()).isPresent()){
+            return appointmentRepository.save(appointment);
+        }
+        else
+            throw new BadRequestException("Patient or dentist do not exist");
+
     }
 
     @Override
@@ -35,17 +47,18 @@ public class AppointmentServiceImpl implements AppointmentService {
     }
 
     @Override
-    public Appointment update(Appointment appointment) throws RuntimeException{
+    public Appointment update(Appointment appointment) throws ResourceNotFoundException {
 
         if (appointment.getId() != null && appointmentRepository.existsById(appointment.getId()))
             return appointmentRepository.save(appointment);
         else
-            throw new RuntimeException();
-
+            throw new ResourceNotFoundException("There is no appointment with id " + appointment.getId());
     }
 
     @Override
-    public void delete(Long id) {
-            appointmentRepository.deleteById(id);
+    public void delete(Long id) throws ResourceNotFoundException {
+        if(findById(id) == null)
+            throw new ResourceNotFoundException("There is no dentist with id " + id);
+        appointmentRepository.deleteById(id);
     }
 }
